+++ bindings_node/test/Conversations.test.ts
@@ -186,9 +186,7 @@ describe('Conversations', () => {
     const user2 = createUser()
     const client1 = await createRegisteredClient(user1)
     const client2 = await createRegisteredClient(user2)
-    const group = await client1
-      .conversations()
-      .findOrCreateDm(user2.account.address)
+    const group = await client1.conversations().createDm(user2.account.address)
     expect(group).toBeDefined()
     expect(group.id()).toBeDefined()
     expect(group.createdAtNs()).toBeTypeOf('number')
@@ -412,9 +410,7 @@ describe('Conversations', () => {
     const group2 = await client2
       .conversations()
       .createGroup([user3.account.address])
-    const group3 = await client4
-      .conversations()
-      .findOrCreateDm(user3.account.address)
+    const group3 = await client4.conversations().createDm(user3.account.address)
 
     await sleep()
 
@@ -436,9 +432,7 @@ describe('Conversations', () => {
     const stream = client3.conversations().streamGroups((err, convo) => {
       groups.push(convo!)
     })
-    const group3 = await client4
-      .conversations()
-      .findOrCreateDm(user3.account.address)
+    const group3 = await client4.conversations().createDm(user3.account.address)
     const group1 = await client1
       .conversations()
       .createGroup([user3.account.address])
@@ -472,9 +466,7 @@ describe('Conversations', () => {
     const group2 = await client2
       .conversations()
       .createGroup([user3.account.address])
-    const group3 = await client4
-      .conversations()
-      .findOrCreateDm(user3.account.address)
+    const group3 = await client4.conversations().createDm(user3.account.address)
 
     await sleep()
 
@@ -494,12 +486,33 @@ describe('Conversations', () => {
     const client4 = await createRegisteredClient(user4)
     await client1.conversations().createGroup([user2.account.address])
     await client1.conversations().createGroup([user3.account.address])
-    await client1.conversations().findOrCreateDm(user4.account.address)
+    await client1.conversations().createDm(user4.account.address)
 
-    let messages: Message[] = []
+    const messages: Message[] = []
     const stream = client1.conversations().streamAllMessages((err, message) => {
       messages.push(message!)
     })
+
+    const messages2: Message[] = []
+    const stream2 = client2
+      .conversations()
+      .streamAllMessages((err, message) => {
+        messages2.push(message!)
+      })
+
+    const messages3: Message[] = []
+    const stream3 = client3
+      .conversations()
+      .streamAllMessages((err, message) => {
+        messages3.push(message!)
+      })
+
+    const messages4: Message[] = []
+    const stream4 = client4
+      .conversations()
+      .streamAllMessages((err, message) => {
+        messages4.push(message!)
+      })
 
     const groups2 = client2.conversations()
     await groups2.sync()
@@ -520,8 +533,17 @@ describe('Conversations', () => {
     await sleep()
 
     stream.end()
+    stream2.end()
+    stream3.end()
+    stream4.end()
     expect(messages.length).toBe(3)
     expect(messages.map((m) => m.id)).toEqual([message1, message2, message3])
+    expect(messages2.length).toBe(1)
+    expect(messages2.map((m) => m.id)).toEqual([message1])
+    expect(messages3.length).toBe(1)
+    expect(messages3.map((m) => m.id)).toEqual([message2])
+    expect(messages4.length).toBe(1)
+    expect(messages4.map((m) => m.id)).toEqual([message3])
   })
 
   it('should only stream group chat messages', async () => {
@@ -535,7 +557,7 @@ describe('Conversations', () => {
     const client4 = await createRegisteredClient(user4)
     await client1.conversations().createGroup([user2.account.address])
     await client1.conversations().createGroup([user3.account.address])
-    await client1.conversations().findOrCreateDm(user4.account.address)
+    await client1.conversations().createDm(user4.account.address)
 
     let messages: Message[] = []
     const stream = client1
@@ -578,7 +600,7 @@ describe('Conversations', () => {
     const client4 = await createRegisteredClient(user4)
     await client1.conversations().createGroup([user2.account.address])
     await client1.conversations().createGroup([user3.account.address])
-    await client1.conversations().findOrCreateDm(user4.account.address)
+    await client1.conversations().createDm(user4.account.address)
 
     let messages: Message[] = []
     const stream = client1