+++ xmtp_mls/src/groups/mod.rs
@@ -1817,9 +1817,20 @@ fn validate_dm_group(
         ));
     }
 
-    // Validate permissions
+    // Validate permissions so no one adds us to a dm that they can unexpectedly add another member to
+    // Note: we don't validate mutable metadata permissions, because they don't affect group membership
     let permissions = extract_group_permissions(mls_group)?;
-    if permissions != GroupMutablePermissions::new(PolicySet::new_dm()) {
+    let expected_permissions = GroupMutablePermissions::new(PolicySet::new_dm());
+
+    if permissions.policies.add_member_policy != expected_permissions.policies.add_member_policy
+        && permissions.policies.remove_member_policy
+            != expected_permissions.policies.remove_member_policy
+        && permissions.policies.add_admin_policy != expected_permissions.policies.add_admin_policy
+        && permissions.policies.remove_admin_policy
+            != expected_permissions.policies.remove_admin_policy
+        && permissions.policies.update_permissions_policy
+            != expected_permissions.policies.update_permissions_policy
+    {
         return Err(GroupError::Generic(
             "Invalid permissions for DM group".to_string(),
         ));